refactor: rename use cases to insert suffix `UseCase`
