+++ src/Application/UseCases/Client/FindAllInactiveClientsUseCase.php
@@ -8,10 +8,11 @@ use JuniorFontenele\LaravelVaultServer\Application\DTOs\Client\ClientResponseDTO
 use JuniorFontenele\LaravelVaultServer\Domains\IAM\Client\Client;
 use JuniorFontenele\LaravelVaultServer\Domains\IAM\Client\Contracts\ClientRepositoryInterface;
 
-class DeleteInactiveClients
+class FindAllInactiveClientsUseCase
 {
-    public function __construct(protected readonly ClientRepositoryInterface $clientRepository)
-    {
+    public function __construct(
+        private readonly ClientRepositoryInterface $clientRepository,
+    ) {
     }
 
     /**
@@ -19,15 +20,11 @@ class DeleteInactiveClients
      */
     public function execute(): array
     {
-        $deletedClients = $this->clientRepository->findAllInactiveClients();
-
-        $this->clientRepository->deleteAllInactiveClients();
-
         return array_map(fn (Client $client) => new ClientResponseDTO(
             clientId: $client->clientId(),
             name: $client->name(),
-            allowedScopes: $client->scopes(),
             description: $client->description(),
-        ), $deletedClients);
+            allowedScopes: $client->scopes()
+        ), $this->clientRepository->findAllInactiveClients());
     }
 }
